Create JSGameProject
This was my first JavaScript Project. I am aware that most of the true developers are more advanced. This is my attempt at something new and exciting. I see how important it is to have fun, and I am proud to create a fun, classic game. I feel this is just the beginning of a healthy addiction I have for developing. 

I also would like to attach a song to some of my post. I feel as if they may encouge others to research inspritaion, and drive you to develop as well.

*Nipsey Hussle - Blue Laces 2
Rest in Power Young King